requirements.txt
pylint
ipykernel
autopep8
matplotlib
kaggle
pandas
seaborn
sklearn
tensorflow-gpu